// apps/trade-pro/src/pages/purchaseReports/purchaseOrderRetailRegister/type.ts
export type TPurchaseOrderSearchCriteria = {
  ItemId?: number;
  ToDocNo?: number;
  FromDocNo?: number;
  ToDate?: Date | string;
  FromDate?: Date | string;
  Status?: number | string;
  SupplierCustomerId?: number;
  IsApproved?: number | string;
};

export type TPurchaseOrderEntry = {
  Id?: number;
  DocNo?: number;
  CityId?: number;
  CommRate?: number;
  PostUser?: number;
  CompanyId?: number;
  EntryUser?: number;
  IsAproved?: boolean;
  PostState?: boolean;
  CommAmount?: number;
  BranchesId?: number;
  ModifyUser?: number;
  ProjectsId?: number;
  BrokeryUom?: number;
  BrokeryRate?: number;
  OrderStatus?: string;
  DeliveryTerm: string;
  CatagorySrNo?: number;
  DeliveryDays?: number;
  OrderDueDays?: number;
  CashFreight?: boolean;
  AttachmentsList?: any;
  OrderSupCustId: number;
  BrokerAgentId?: number;
  BrokeryAmount?: number;
  OrderTaxable?: boolean;
  DocDate: Date | string;
  PaymentTermsId: number;
  DocumentTypeId?: number;
  OrganizationId?: number;
  CreditFreight?: boolean;
  FinancialYearId?: number;
  OrderCatagoryId?: number;
  RefrenenceParty?: number;
  EntryDate?: Date | string;
  OrderType?: string | null;
  ModifyDate?: Date | string;
  OrderDueDate: Date | string;
  BrokeryType?: string | null;
  UomScheduleIdCmRate?: number;
  BrokerAgentSupCustId?: number;
  RemarksHeader?: string | null;
  SupplierRefNo?: number | null;
  CommissionType?: string | null;
  OrderExpiryDate?: Date | string;
  PostDate?: Date | string | null;
  DeliveryRemarks?: string | null;
  PurchaseorderEmptyBagsList?: any;
  DeliveryStartDate?: Date | string;
  CommissionRemarks?: string | null;
  PurchaseOrderLabDeductionList?: any;
  SupplierCustomerIdStockParty?: number;
  PurchaseOrderExpensesChargeToProductsList?: any;
  purchaseOrderDetailList: TPurchaseOrderDetailEntry[];
};

export type TPurchaseOrderDetailEntry = {
  Id?: number;
  LineId?: number;
  UOMCode?: string;
  RateUom?: string;
  BagPrice?: number;
  ItemName?: string;
  BagWeight?: number;
  TaxAmount?: number;
  TaxNameId?: number;
  RetailRate?: number;
  TaxPercent?: number;
  TotalAmount?: number;
  Crop?: string | null;
  ActionTypeId?: number;
  Amount?: number | null;
  OrderItemRate?: number;
  OrderItemUOMId?: number;
  TaxName?: string | null;
  UOMDescription?: number;
  JobLotId?: null | number;
  PurchaseOrderId?: number;
  CityArea?: string | null;
  Moisture?: string | null;
  NetWeight?: number | null;
  JobLotDescription?: string;
  OrderItemId?: null | number;
  OrderItemRateUOMId?: number;
  LabSampleNo?: number | null;
  OrderItemQty?: number | null;
  OrderRemarks?: string | null;
  TaxableStatus?: string | null;
  EquivalentRate?: number | null;
  AmountCalcType?: string | null;
  InvLabSampleAnalysisHeaderId?: number;
};